requirements.txt
selenium>=4.0
requests>=2.28
webdriver-manager>=4.0
google-generativeai>=0.3.0
pandas>=2.0
openpyxl>=3.1
